requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
python-multipart==0.0.6
PyMuPDF==1.22.5
langchain==0.0.312
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.11.0
tqdm==4.65.0
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0